requirements.txt
pandas
fastapi
uvicorn
pyarrow.parquet